user-story-tests/32-markdown-component-with-page-scope.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Markdown component with page @scope</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <script type="module">
        import { MDBlockWithUserSelectablePageStyles } from "./components/MDBlockWithUserSelectablePageStyles.js";
        customElements.define(
            "md-block-with-user-selectable-page-styles",
            MDBlockWithUserSelectablePageStyles
        );
    </script>

    <style>
        @layer markdown-scope-styles {
            @scope (article) to (li) {
                ul {
                    border: thick dashed red;
                }
            }
        }
    </style>
    <h2>H2 outside shadow tree (unstyled)</h2>
    <md-block-with-user-selectable-page-styles>
        <script type="text/markdown">
## H2 from markdown (unstyled)

- list item from markdown (thick dashed red) (from page @scope)

### H3 from markdown (thick solid black) (from component defaults)
    </script>
        <template shadowrootmode="open">
            <style>
                @layer inherit.markdown-scope-styles.as.markdown-scope-styles, md-block-default-styles, markdown-scope-styles;
            </style>
        </template>
    </md-block-with-user-selectable-page-styles>

    <script type="module">
        import { adoptStylesFromLayerStatementRule } from "../AdoptStyles.js";

        window.onload = function () {

            test(function () {

                assert_true(true, "always true, just to show the test is running");

            }, "Markdown component with page @scope");

        }
    </script>

    <div id="log"></div>

</body>

</html>
